pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel", "build"]
build-backend = "setuptools.build_meta"

[project]
name = 'mdar'
version = '0.1.0'
description = "Markdown Archiver (mdar) - A tool to archive text files in markdown."
authors = [
    { name = "Philip Dye", email = "phdye@acm.org" },
]
requires-python = ">=3.8"
license = { text = "MIT" }
readme = 'README.md'
dependencies = [
    "docopt-ng>=0.9.0",
    "toml>=0.10.2",
]
optional-dependencies = { 'testing' = [
        'pytest>=6.0',
        'pytest-cov>=2.5.1',
        'flake8>=3.8',
    ] }

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pdm]
